refactor: Notification 도메인 관리자/사용자 폴더 분리
