+++ src/main/java/kr/tennispark/common/exception/GlobalExceptionHandler.java
@@ -11,8 +11,8 @@ import kr.tennispark.common.exception.base.NotFoundException;
 import kr.tennispark.common.exception.base.UnsupportedTypeException;
 import kr.tennispark.common.utils.ApiUtils;
 import kr.tennispark.common.utils.ApiUtils.ApiResult;
-import kr.tennispark.notification.application.exception.FcmMessageSendFailureException;
-import kr.tennispark.notification.infrastructure.exception.FirebaseInitializationException;
+import kr.tennispark.notification.admin.application.exception.FcmMessageSendFailureException;
+import kr.tennispark.notification.admin.infrastructure.exception.FirebaseInitializationException;
 import kr.tennispark.qr.application.exception.BadQrCreateException;
 import kr.tennispark.qr.application.exception.FailToCreateQRPayloadException;
 import lombok.extern.slf4j.Slf4j;
