+++ src/main/java/kr/tennispark/common/scheduler/SendActivityNotificationScheduler.java
@@ -8,11 +8,11 @@ import java.util.Map;
 import kr.tennispark.activity.admin.infrastructure.repository.AdminActivityApplicationRepository;
 import kr.tennispark.activity.common.domain.Activity;
 import kr.tennispark.members.common.domain.entity.Member;
-import kr.tennispark.notification.application.NotificationMessageFactory;
-import kr.tennispark.notification.application.NotificationPublisher;
-import kr.tennispark.notification.domain.entity.NotificationSchedule;
-import kr.tennispark.notification.domain.entity.enums.NotificationCategory;
-import kr.tennispark.notification.infrastructure.NotificationScheduleRepository;
+import kr.tennispark.notification.admin.application.NotificationMessageFactory;
+import kr.tennispark.notification.admin.application.NotificationPublisher;
+import kr.tennispark.notification.admin.infrastructure.NotificationScheduleRepository;
+import kr.tennispark.notification.common.domain.entity.NotificationSchedule;
+import kr.tennispark.notification.common.domain.entity.enums.NotificationCategory;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
