+++ src/main/java/kr/tennispark/common/scheduler/SendRecruitReminderScheduler.java
@@ -5,8 +5,8 @@ import java.util.List;
 import kr.tennispark.activity.admin.infrastructure.repository.ActivityRepository;
 import kr.tennispark.activity.common.domain.Activity;
 import kr.tennispark.activity.common.domain.vo.WeekPeriod;
-import kr.tennispark.notification.application.NotificationMessageFactory;
-import kr.tennispark.notification.application.NotificationPublisher;
+import kr.tennispark.notification.admin.application.NotificationMessageFactory;
+import kr.tennispark.notification.admin.application.NotificationPublisher;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
