+++ src/main/java/kr/tennispark/notification/admin/infrastructure/NotificationScheduleRepository.java
@@ -1,12 +1,12 @@
-package kr.tennispark.notification.infrastructure;
+package kr.tennispark.notification.admin.infrastructure;
 
 import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.List;
 import kr.tennispark.activity.common.domain.Activity;
 import kr.tennispark.members.common.domain.entity.Member;
-import kr.tennispark.notification.domain.entity.NotificationSchedule;
-import kr.tennispark.notification.domain.entity.enums.NotificationType;
+import kr.tennispark.notification.common.domain.entity.NotificationSchedule;
+import kr.tennispark.notification.common.domain.entity.enums.NotificationType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
