+++ src/main/java/kr/tennispark/notification/admin/infrastructure/exception/FirebaseInitializationException.java
@@ -1,4 +1,4 @@
-package kr.tennispark.notification.infrastructure.exception;
+package kr.tennispark.notification.admin.infrastructure.exception;
 
 public class FirebaseInitializationException extends RuntimeException {
     private static final String DEFAULT_MESSAGE = "Firebase 초기화에 실패하였습니다.";
