+++ src/main/java/kr/tennispark/notification/admin/presentation/request/SendMessageRequestDTO.java
@@ -1,4 +1,4 @@
-package kr.tennispark.notification.presentation.request;
+package kr.tennispark.notification.admin.presentation.request;
 
 public record SendMessageRequestDTO(String title, String content) {
 }
