+++ src/main/java/kr/tennispark/notification/common/domain/entity/Notification.java
@@ -1,4 +1,4 @@
-package kr.tennispark.notification.domain.entity;
+package kr.tennispark.notification.common.domain.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -11,7 +11,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import kr.tennispark.common.domain.BaseEntity;
 import kr.tennispark.members.common.domain.entity.Member;
-import kr.tennispark.notification.domain.entity.enums.NotificationCategory;
+import kr.tennispark.notification.common.domain.entity.enums.NotificationCategory;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
